Fix circular pt2
